Rework ECS framework
- Add queue systems when adding/removing components and systems
- Add GlobalData to ECSManager, used to store data that is not specific
  to entities
- Add View class, used to for loop all entities that contains the given
  Components
- Rework how signaling works

+++ CopiumEngine/src/copium/core/SwapChain.cpp
@@ -430,7 +430,7 @@ namespace Copium
     int width, height;
     glfwGetFramebufferSize(window, &width, &height);
 
-    VkExtent2D extent{width, height};
+    VkExtent2D extent{static_cast<uint32_t>(width), static_cast<uint32_t>(height)};
     extent.width = std::clamp(extent.width, capabilities.minImageExtent.width, capabilities.maxImageExtent.width);
     extent.height = std::clamp(extent.height, capabilities.minImageExtent.height, capabilities.maxImageExtent.height);
     return extent;
